feat: Trade

+++ pkg/events/order/order_base.go
@@ -231,7 +231,7 @@ func (b *Base) Time() int64 {
 }
 
 func (b *Base) SetTime(value int64) {
-	b.index = (int64(timeutil.GetSecondsFromTime(value)) << 32) |
+	b.index = (timeutil.GetSecondsFromTime(value) << 32) |
 		int64(timeutil.GetMillisFromTime(value)<<22) |
 		b.Sequence()
 }
