+++ pkg/events/timeandsale/time_and_sale.go
@@ -184,7 +184,7 @@ func (t *TimeAndSale) Time() int64 {
 }
 
 func (t *TimeAndSale) SetTime(value int64) {
-	t.index = (int64(timeutil.GetSecondsFromTime(value)) << 32) |
+	t.index = (timeutil.GetSecondsFromTime(value) << 32) |
 		int64(timeutil.GetMillisFromTime(value)<<22) |
 		t.Sequence()
 }
